@@ -1,23 +1,30 @@
 <?php
+
+use App\Http\Controllers\CartController;
+use App\Http\Controllers\CustomerController;
+use App\Http\Controllers\HomeController;
+use App\Http\Controllers\OrderController;
+use App\Http\Controllers\ProductController;
+use App\Http\Controllers\SettingController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
-    return view('welcome');
+    return redirect('/admin');
 });
 
 Auth::routes();
 
 Route::prefix('admin')->middleware('auth')->group(function () {
-    Route::get('/', 'HomeController@index')->name('home');
-    Route::get('/settings', 'SettingController@index')->name('settings.index');
-    Route::post('/settings', 'SettingController@store')->name('settings.store');
-    Route::resource('products', 'ProductController');
-    Route::resource('customers', 'CustomerController');
-    Route::resource('orders', 'OrderController');
+    Route::get('/', [HomeController::class, 'index'])->name('home');
+    Route::get('/settings', [SettingController::class, 'index'])->name('settings.index');
+    Route::post('/settings', [SettingController::class, 'store'])->name('settings.store');
+    Route::resource('products', ProductController::class);
+    Route::resource('customers', CustomerController::class);
+    Route::resource('orders', OrderController::class);
 
-    Route::get('/cart', 'CartController@index')->name('cart.index');
-    Route::post('/cart', 'CartController@store')->name('cart.store');
-    Route::post('/cart/change-qty', 'CartController@changeQty');
-    Route::delete('/cart/delete', 'CartController@delete');
-    Route::delete('/cart/empty', 'CartController@empty');
+    Route::get('/cart', [CartController::class, 'index'])->name('cart.index');
+    Route::post('/cart', [CartController::class, 'store'])->name('cart.store');
+    Route::post('/cart/change-qty', [CartController::class, 'changeQty']);
+    Route::delete('/cart/delete', [CartController::class, 'delete']);
+    Route::delete('/cart/empty', [CartController::class, 'empty']);
 });
