@@ -3,6 +3,7 @@
 namespace App\Providers;
 
 use App\Models\Setting;
+use Illuminate\Pagination\Paginator;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -24,17 +25,21 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        // 'key' => 'value'
-        $settings = Setting::all('key', 'value')
-            ->keyBy('key')
-            ->transform(function ($setting) {
-                return $setting->value;
-            })
-            ->toArray();
-        config([
-           'settings' => $settings
-        ]);
+        if (! $this->app->runningInConsole()) {
+            // 'key' => 'value'
+            $settings = Setting::all('key', 'value')
+                ->keyBy('key')
+                ->transform(function ($setting) {
+                    return $setting->value;
+                })
+                ->toArray();
+            config([
+               'settings' => $settings
+            ]);
 
-        config(['app.name' => config('settings.app_name')]);
+            config(['app.name' => config('settings.app_name')]);
+        }
+
+        Paginator::useBootstrap();
     }
 }
